+++ common/src/main/java/com/stevekung/fishofthieves/registry/FOTBlocks.java
@@ -112,7 +112,7 @@ public class FOTBlocks
     public static final Block HANGING_MANGO_FRUIT = new HangingMangoFruitBlock(BlockBehaviour.Properties.of().offsetType(BlockBehaviour.OffsetType.XYZ).dynamicShape().randomTicks().noOcclusion().sound(SoundType.WOOD).instabreak().pushReaction(PushReaction.DESTROY));
     public static final Block TALL_POMEGRANATE_PLANT = new TallPomegranatePlantBlock(BlockBehaviour.Properties.of().mapColor(MapColor.PLANT).instabreak().sound(SoundType.AZALEA).ignitedByLava().pushReaction(PushReaction.DESTROY));
     public static final Block PRISMARIZED_LOG = new FOTRotatedPillarBlock(BlockBehaviour.Properties.of().lightLevel(blockState -> 3).mapColor(MapColor.COLOR_LIGHT_BLUE).instrument(NoteBlockInstrument.BASS).strength(2.0F).sound(SoundType.NETHER_WOOD).ignitedByLava().emissiveRendering(FOTBlocks::always));
-    public static final Block BUDDING_GUARDIAN_FRUIT = new BuddingGuardianFruitBlock(BlockBehaviour.Properties.of().lightLevel(blockState -> 3).mapColor(MapColor.COLOR_LIGHT_BLUE).randomTicks().strength(2.0F).instrument(NoteBlockInstrument.BASS).sound(SoundType.NETHER_WOOD).pushReaction(PushReaction.DESTROY).ignitedByLava().emissiveRendering(FOTBlocks::always));
+    public static final Block BUDDING_PRISMARIZED_LOG = new BuddingPrismarizedLogBlock(BlockBehaviour.Properties.of().lightLevel(blockState -> 3).mapColor(MapColor.COLOR_LIGHT_BLUE).randomTicks().strength(2.0F).instrument(NoteBlockInstrument.BASS).sound(SoundType.NETHER_WOOD).pushReaction(PushReaction.DESTROY).ignitedByLava().emissiveRendering(FOTBlocks::always));
     public static final Block GUARDIAN_FRUIT = new GuardianFruitBlock(BlockBehaviour.Properties.of().noOcclusion().sound(SoundType.WOOD).instabreak().pushReaction(PushReaction.DESTROY).isSuffocating(FOTBlocks::never).isRedstoneConductor(FOTBlocks::never).isValidSpawn(FOTBlocks::never));
 
     public static Block PINK_PLUMERIA;
@@ -247,7 +247,7 @@ public class FOTBlocks
         registerNoItem("hanging_mango_fruit", HANGING_MANGO_FRUIT);
         registerNoItem("tall_pomegranate_plant", TALL_POMEGRANATE_PLANT);
         register("prismarized_log", PRISMARIZED_LOG);
-        register("budding_guardian_fruit", BUDDING_GUARDIAN_FRUIT);
+        register("budding_prismarized_log", BUDDING_PRISMARIZED_LOG);
         registerNoItem("guardian_fruit", GUARDIAN_FRUIT);
 
         register("coconut_planks", COCONUT_PLANKS);